// src/app/Models/PersonalProposal/consumer-durables-loan.ts
export class ConsumerDuarablesLoan {
    ID: number;
    CLIENT_ID: number;
    PROPOSAL_ID: number;
    DURABLE_NAME: string;
    OTHER_DURABLE_NAME: string;
    DURABLE_TYPE: string ='R';
    IS_GIVEN_ADVANCE:boolean =false;
    ADVANCE_AMOUNT: number;
    DEALER_NAME: string;
    IS_QUOTATION_TAKEN:boolean =false;
    QUOTATION_AMOUNT: number;
    LIFE_OF_PRODUCT: number;
    BRAND_NAME:string
}